fetch coins market data moved to main screen instead of component
